front page moved form comonents to page folder
