+++ src/App.jsx
@@ -1,6 +1,6 @@
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom'
 import './App.css';
-import FrontPage from './components/frontPage/FrontPage';
+import FrontPage from './pages/frontPage/FrontPage';
 import RegisterPage from './pages/registerPage/RegisterPage'; 
 import LoginPage from './pages/loginPage/LoginPage';
